ZEPPELIN-3208. Use interpreter setting's name as its id
### What is this PR for?
Interpreter setting's name is unique, so it is not necessary to use id to identify it. For now we use a random string to represent such id which is not easy to read especially when you read log for diagnose. But for backward compatibility, I will keep id and set it as the same as interpreter setting's name

### What type of PR is it?
[Improvement]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-3208

### How should this be tested?
* CI pass

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #2778 from zjffdu/ZEPPELIN-3208 and squashes the following commits:

56a33c2 [Jeff Zhang] ZEPPELIN-3208. Use interpreter setting's name as its id

(cherry picked from commit b3ba458bd22f4320162cbb4baf3bab48a6844b30)