[ZEPPELIN-3626] Cluster server and client module design
The cluster management server uses the Raft algorithm library Atomix to form a service cluster with consistent service status in the Zeppelin cluster. The cluster management client connects to the cluster management server for metadata operations of services and processes. The cluster monitoring module checks if each Zeppelin-Server and Zeppelin Interpreter process in the cluster is active.

[Feature]

* [x] add cluster manager server
* [x] add cluster manager client
* [x] add cluster monitor
* [x] add cluster create Interpreter thrift interface
* [x] add InterpreterFactoryInterface
* [x] add the function call in the ZeppelinServer startup cluster

* https://issues.apache.org/jira/browse/ZEPPELIN-3626

CI pass

* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? [Design document](https://docs.google.com/document/d/1a8QLSyR3M5AhlG1GIYuDTj6bwazeuVDKCRRBm-Qa3Bw/edit#heading=h.qbcgqhd0wwh8)